Added explicit cast to convert 32-bit tick count to a 64-bit value

+++ src/platform/thread_win.cpp
@@ -52,7 +52,7 @@ void msleep(unsigned long timeout)
 
 uint64_t msecs(void)
 {
-   return GetTickCount();
+   return (uint64_t) GetTickCount();
 }
 
 unsigned long thread_id(void)
